Moved web worker to its own feature
