+++ next.config.mjs
@@ -1,13 +1,14 @@
-import withSerwistInit from "@serwist/next";
+// import withSerwistInit from "@serwist/next";
 
-const withSerwist = withSerwistInit({
-  cacheOnFrontEndNav: true,
-  swSrc: "src/app/service-worker.ts",
-  swDest: "public/sw.js",
-  injectionPoint: "self.__SW_MANIFEST",
-});
+// const withSerwist = withSerwistInit({
+//   cacheOnFrontEndNav: true,
+//   swSrc: "src/app/service-worker.ts",
+//   swDest: "public/sw.js",
+//   injectionPoint: "self.__SW_MANIFEST",
+// });
 
 /** @type {import("next").NextConfig} */
 const nextConfig = {};
 
-export default withSerwist(nextConfig);
+// export default withSerwist(nextConfig);
+export default nextConfig;
