+++ src/types/time.ts
@@ -1,0 +1,2 @@
+/** A helper type to denote that a number in code is meant to be milliseconds */
+export type Milliseconds = number;
